+++ services/web/server/src/simcore_service_webserver/projects/_handlers_crud.py
@@ -23,7 +23,6 @@ from models_library.rest_pagination import (
     Page,
 )
 from models_library.rest_pagination_utils import paginate_data
-from models_library.users import UserID
 from models_library.utils.fastapi_encoders import jsonable_encoder
 from pydantic import BaseModel, Extra, Field, NonNegativeInt, validator
 from servicelib.aiohttp.long_running_tasks.server import start_long_running_task
@@ -38,10 +37,8 @@ from servicelib.common_headers import (
 )
 from servicelib.json_serialization import json_dumps
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
-from servicelib.request_keys import RQT_USERID_KEY
 from servicelib.rest_constants import RESPONSE_MODEL_POLICY
 
-from .._constants import RQ_PRODUCT_KEY
 from .._meta import api_version_prefix as VTAG
 from ..catalog.client import get_services_for_user_in_product
 from ..director_v2 import api
@@ -51,6 +48,7 @@ from ..security.api import check_permission
 from ..security.decorators import permission_required
 from ..users.api import get_user_name
 from . import _crud_create_utils, _crud_read_utils, projects_api
+from ._common_models import ProjectPathParams, RequestContext
 from ._crud_read_utils import OrderDirection, ProjectListFilters, ProjectOrderBy
 from ._permalink_api import update_or_pop_permalink_in_project
 from .db import ProjectDBAPI
@@ -76,22 +74,9 @@ from .utils import (
 RQ_REQUESTED_REPO_PROJECT_UUID_KEY = f"{__name__}.RQT_REQUESTED_REPO_PROJECT_UUID_KEY"
 
 
-log = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)
 
 routes = web.RouteTableDef()
-
-
-class RequestContext(BaseModel):
-    user_id: UserID = Field(..., alias=RQT_USERID_KEY)  # type: ignore[pydantic-alias]
-    product_name: str = Field(..., alias=RQ_PRODUCT_KEY)  # type: ignore[pydantic-alias]
-
-
-class ProjectPathParams(BaseModel):
-    project_id: ProjectID
-
-    class Config:
-        allow_population_by_field_name = True
-        extra = Extra.forbid
 
 
 #
@@ -105,15 +90,15 @@ class _ProjectCreateParams(BaseModel):
         description="Option to create a project from existing template or study: from_study={study_uuid}",
     )
     as_template: bool = Field(
-        False,
+        default=False,
         description="Option to create a template from existing project: as_template=true",
     )
     copy_data: bool = Field(
-        True,
+        default=True,
         description="Option to copy data when creating from an existing template or as a template, defaults to True",
     )
     hidden: bool = Field(
-        False,
+        default=False,
         description="Enables/disables hidden flag. Hidden projects are by default unlisted",
     )
 
@@ -227,9 +212,8 @@ class _ProjectListParams(BaseModel):
                 if field_info[1] == OrderDirection.DESC.value:
                     direction = OrderDirection.DESC
                 else:
-                    raise ValueError(
-                        "Field direction in the order_by parameter must contain either 'desc' direction or empty value for 'asc' direction."
-                    )
+                    msg = "Field direction in the order_by parameter must contain either 'desc' direction or empty value for 'asc' direction."
+                    raise ValueError(msg)
 
             parse_fields_with_direction.append(
                 ProjectOrderBy(field=field_name, direction=direction)
